Basic types

The package has a few types to represent objects in a Virasoro CFT:

CentralCharges

ConformalDimensions

Fields

Central charges

We parametrise the central charge of our theories in terms of variables c, B, b or β related by

\[c = 13 + 6B + 6 B^{-1} \quad , \quad B = b^2 = -\beta^2, \quad B = \frac{c-13 \pm \sqrt{(c-1)(c-25)}}{12}\]

By convention we keep $\beta = ib$. In $O(n)$ and $U(n)$ models the central charge is related to $n$ via

\[n = - 2 \cos(\pi \beta^2)\]

The program allows to conveniently create central charges from any of these four parameters, and to retrieve the value of any parameter:

BootstrapVirasoro.CentralCharge — Type

CentralCharge(param = value)

Type representing a central charge, with precision given by the type of value. The supported param are c, β, b, B.

Examples

julia> c = CentralCharge(β=big"0.1"+big"0.2"*im); c.β ≈ big"0.1" + big"0.2"*im
true

julia> c = CentralCharge(c = 0.7);

julia> c.b ≈ -0.0 + 0.894427190999916im
true

julia> c.n ≈ 1.6180339887498953 + 0.0im
true

Conformal Dimensions

We parametrise the conformal dimensions in terms of variables $\Delta, P, p, \delta$, related by

\[\Delta = \frac{c-1}{24} + \delta \quad , \quad \delta = P^2 = -p^2\]

The variable $P$ is called the momentum, and $\Delta$ is the eigenvalue of the Virasoro generator $L_0$. By convention, we always keep $P=ip$. Moreover, we introduce the following parametrisation of dimensions in terms of Kac indices $r, s$:

\[P_{(r,s)}=\frac{1}{2}(\beta r - \beta^{-1}s)\]

where $r,s$ are arbitrary numbers. This convention is different from the one in Sylvain\'s code, but similar to our more recent conventions, such as in Sylvain's review on solvable CFTs.

The program lets us define these objects and access the various parametrisations:

BootstrapVirasoro.ConformalDimension — Type

ConformalDimension(c; param=value, r=0, s=0)

Type representing a conformal dimension, with precision given by the central charge. The supported parameters are Δ, δ, P, p, or the Kac indices r, s.

Examples

julia> c = CentralCharge(β = 0.3im);

julia> d1 = ConformalDimension(c, δ=0.5); d1.Δ ≈ 3.800277777777777 + 0.0im
true

julia> d1.P ≈ 0.7071067811865476
true

julia> d2 = ConformalDimension(c, r=3//2, s=2//3)
ConformalDimension{ComplexF64} with Kac indices r = 3//2, s = 2//3

BootstrapVirasoro.P_rs — Function

    P_rs(r, s, β)
    P_rs(r, s, c)

$P_{(r, s)} = 1/2 * (β r - s / β)$.

ConformalDimensions can be shifted by units of $\beta^{-1}/2$:

BootstrapVirasoro.shift — Method

    shift(d, i)

Shift a conformal dimension by s -> s+i or P -> P + i/β

Fields

For our purposes, a field is the data of left and right conformal dimensions. We denote $V_{(r, s)}$ a field with left and right dimensions $(\Delta_{(r, s)}, \Delta_{(r, -s)})$.

The program exposes a Field struct and convenient constructors:

BootstrapVirasoro.Field — Type

Field(c, parameter=value; r, s, diagonal=false)

Type for representing a non-chiral field.

keyword arguments:

r::Rational,s::Rational. By convention $V_{(r,s)}$ has left and right momenta $(P_{(r,s)}, P_{(r,-s)})$.

diagonal::Bool: set to true to get a diagonal field;

Examples

julia> c = CentralCharge(β = big"0.5");

julia> V = Field(c, r=0, s=1)
Diagonal Field{Complex{BigFloat}}, dim = Δ_{0, 1}

julia> V.Δ
(0.4375 + 0.0im, 0.4375 + 0.0im)

julia> V.P[:left] ≈ 1
true

julia> V.p[:right] ≈ -im
true

julia> V2 = Field(c, :P, 0.42, diagonal=true); isdiagonal(V2)
true

BootstrapVirasoro.isdiagonal — Function

    isdiagonal(V)::Bool

Whether V is a diagonal field.

BootstrapVirasoro.isdegenerate — Function

    isdegenerate(V)::Bool

Whether V is a degenerate field.

BootstrapVirasoro.shift — Method

    shift(V, i)

Shift the field:

s -> s+i if !isdiagonal(V)

P -> P+i/(2β) if isdiagonal(V)

Spectra

The program exposes two types to deal with CFT spectra.

The type Spectrum simply contains a list of conformal dimensions or fields, with conformal dimensions bounded by some $\Delta_{\mathrm{max}}$.

The type ChannelSpectrum contains all the data needed to evaluate blocks corresponding to a correlation and a list of channel fields in a given channel.

BootstrapVirasoro.Spectrum — Type

Spectrum(fields, Δmax; interchiral=false)

Abstract type for representing a CFT spectrum.

examples

julia> c = CentralCharge(β=1/(big"0.8"+big"0.1"*im));

julia> V1 = Field(c, r=2, s=0);

julia> co = Correlation(V1, Δmax=10.);

julia> fields = [Field(c, r=r, s=s) for r in 2:30 for s in -1+1//r:1//r:1 if r*s% 1 == 0];

julia> s = Spectrum(fields, 10.0)
Non-diagonal:
(2, -1//2), (2, 0), (2, 1//2), (2, 1)
(3, -2//3), (3, -1//3), (3, 0), (3, 1//3), (3, 2//3), (3, 1)

julia> add!(s, Field(c, r=1, s=0)); s
Non-diagonal:
(1, 0)
(2, -1//2), (2, 0), (2, 1//2), (2, 1)
(3, -2//3), (3, -1//3), (3, 0), (3, 1//3), (3, 2//3), (3, 1)

BootstrapVirasoro.ChannelSpectrum — Type

    Holds all the data for evaluating blocks in a channel.