Computing conformal blocks

The library uses the Zamolodchikov recursion to compute four-point conformal blocks on the sphere, and one-point conformal blocks on the torus. Because it is common in bootstrap computations to have to evaluate a given block at hundreds of positions, the library first creates a Block object which stores the coefficients of the block's series expansion, which can then be evaluated.

The library provides a hierarchy of conformal block types:

Block  (Abstract Type)
├─ ChiralBlock
├─ NonChiralBlock  (Abstract Type)
│  ├─ Factorised block
│  └─ LogBlock
└─ LinearCombinationBlock

Defining and computing blocks requires defining a central charge, and conformal dimensions of the external and channel fields.

BootstrapVirasoro.CentralCharge — Type

    CentralCharge(; β, c, B, b)

Create a CentralCharge object from one of the parameters $β, c = 13 - 6 β^2 - 6 β^{-2}, b = i β, B = b^2$.

Aliased to CC.

Examples

c1 = CentralCharge(β=0.5+0.3im)
c2 = CC(c=big"0.4")
c2.c ≈ big"0.4" # access any of the parameters

BootstrapVirasoro.ConformalDimension — Type

    ConformalDimension(c::CC; Δ, δ, P, p, r, s)

Create a ConformalDimension object from one of the parameters $P, p = iP, δ = P^2, Δ = δ + \frac{c-1}{24}$ or from Kac indices r, s ($P_{(r, s)} = \frac{1}{2}(rβ - sβ^{-1})$). The index r must be integer or rational, s can be arbitrary.

Aliased to CD.

Examples

c = CC(β = big"0.3"+big"0.1"*im)
d1 = CD(c, P=big"0.5"*im)
d2 = CD(c, r=2, s=1//2)
d3 = CD(c, r=0, s=big"0.1")
d2.r == 2 # true

BootstrapVirasoro.Field — Type

    Field(c::CC; Δ, δ, P, p, r, s, diagonal)

Create a Field object from one of the parameters P, p, δ, Δ (see ConformalDimension) or from Kac indices r, s. The index r must be integer or rational, s can be arbitrary. Left and right dimensions are accessed with [:left], [:right].

Examples

c = CC(β = big"0.3"+big"0.1"*im)
V1 = Field(c, r=2, s=1//2) # non-diagonal, (Δ(2, 1//2), Δ(2, -1//2))
V2 = Field(c, P=0.5)       # diagonal
V3 = Field(c, r=1, s=2, diagonal=true) # degenerate
V3[:left].Δ == V3[:right].Δ # true

The parameters of the external fields are stored in a Correlation object, that also stores the residues $R_{m, n}$ for each channel, and their sums

\[C^N_{m,n} = R_{m,n}\left(\delta_{N-mn,0} + \sum_{m'n'\leq N-mn} \frac{C^{N-mn}_{m',n'}}{\delta_{(m,-n)}-\delta_{(m',n')}} \right)\]

used in the Zamolodchikov recursion.

Correlation  (Abstract Type)
├─ Correlation1  (Abstract Type)
│  ├─ ChiralCorrelation1
│  └─ NonChiralCorrelation1
└─ Correlation4  (Abstract Type)
   ├─ ChiralCorrelation4
   └─ NonChiralCorrelation4

They are all created via a common interface

BootstrapVirasoro.Correlation — Type

    Correlation(fields, Δmax)

Create a Correlation object. The fields can be passed as single arguments or as tuples, and there can be 1 or 4 of them. The Correlation object stores residues $R_{m, n}$ and $C^N_{m, n}$ up to Δmax.

Aliased to Corr, Co.

Examples

c = CC(β = 0.5)
V = Field(c, r=2, s=1//2)
co1 = Correlation(V, V, V, V, 10) # NonChiralCorrelation4
co2 = Corr(V, 10) # NonChiralCorrelation1
co3 = Corr(Tuple(V[:left] for _ in 1:4), 20) # ChiralCorrelation4
co3.fields[1] == V[:left]