Basic types

The package has a few types to represent objects in a Virasoro CFT:

CentralCharges

ConformalDimensions

Fields

Central charges

We parametrise the central charge of our theories in terms of variables c, B, b or \beta related by

\[c = 13 + 6B + 6 B^{-1} \quad , \quad B = b^2 = -\beta^2, \quad B = \frac{c-13 \pm \sqrt{(c-1)(c-25)}}{12}\]

By convention we keep $\beta = ib$. In $O(n)$ and $U(n)$ models the central charge is related to $n$ via

\[n = - 2 \cos(\pi \beta^2)\]

The program allows to conveniently create central charges from any of these four parameters, and to retrieve the value of any parameter:

BootstrapVirasoro.CentralCharge — Type

CentralCharge{T}

Type representing a central charge. T is expected to be a real or complex number, of standard or arbitrary precision. The supported parameters are c, β, b, B.

Examples

julia> c = CentralCharge(c = 0.7)
c = 0.7000000000000011 + 0.0im, β = -0.894427190999916 - 0.0im

julia> c.b
-0.0 + 0.894427190999916im

julia> c.n
1.6180339887498953 + 0.0im

BootstrapVirasoro.CentralCharge — Method

CentralCharge(parameter = value)

Constructor function for the CentralCharge type.

Given one of the four parameters c, b, β, B and its value, creates an object CentralCharge{T}.

Example

julia> setprecision(BigFloat, 20, base=10);

julia> CentralCharge(B = 0.5)
c = 27.999999999999996 + 0.0im, β = 0.0 - 0.7071067811865476im

julia> CentralCharge(β = 0.7)
c = -2.184897959183676 + 0.0im, β = -0.7 - 0.0im

julia> CentralCharge(c = big"0.1" + 0.2im)
c = 0.09999999999999999991326 + 0.2000000000000000111158im, β = -0.8237591041762989640376 - 0.01729590504934815486866im

Conformal Dimensions

We parametrise the conformal dimensions in terms of variables $\Delta, P, p, \delta$, related by

\[\Delta = \frac{c-1}{24} + \delta \quad , \quad \delta = P^2 = -p^2\]

The variable $P$ is called the momentum, and $\Delta$ is the eigenvalue of the dilation operator. By convention, we always keep $P=ip$. Moreover, we introduce the following parametrisation of dimensions in terms of Kac indices $r, s$:

\[P_{(r,s)}=\frac{1}{2}(\beta r - \beta^{-1}s)\]

where $r,s$ are arbitrary numbers. This convention is different from the one in Sylvain\'s code, but similar to our more recent conventions, such as in Sylvain's review on solvable CFTs.

The program lets us define these objects and access the various parametrisations:

BootstrapVirasoro.ConformalDimension — Type

ConformalDimension{T}

Type for encoding a conformal dimension. The supported parameters are Δ, δ, P, p, w, or the Kac indices r and s.

Examples

julia> c = CentralCharge(:c, 0.5);

julia> d1 = ConformalDimension(c, P = 1.2+0.1im)
ConformalDimension{ComplexF64} with
Δ = 1.4091666666666667 + 0.24im, P = 1.2 + 0.1im

julia> d2 = ConformalDimension(c, r=2, s=3//2)
ConformalDimension{ComplexF64} with Kac indices r = 2//1, s = 3//2

julia> d1.P
1.2 + 0.1im

julia> d2.Δ
-0.020833333333333332 + 0.0im

julia> d1 + d2
ConformalDimension{ComplexF64} with
Δ = 1.3883333333333332 + 0.24im, P = 1.1913489935345947 + 0.10072615216131914im

julia> d1.w
2.05874441299789 - 0.06772184182090507im

BootstrapVirasoro.ConformalDimension — Method

ConformalDimension(c, parameter, value; Kac=false, r=0, s=0)

Constructor function for the ConformalDimension type.

Examples

julia> c = CentralCharge(β = 0.3im)
c = 80.20666666666665 + 0.0im, β = 0.0 - 0.3im

julia> d1 = ConformalDimension(c, δ = 0.5)
ConformalDimension{ComplexF64} with
Δ = 3.800277777777777 + 0.0im, P = 0.7071067811865476

julia> d2 = ConformalDimension(c, r=3//2, s=2//3)
ConformalDimension{ComplexF64} with Kac indices r = 3//2, s = 2//3

Fields

For our purposes, a field is the data of left and right conformal dimensions. We denote $V_{(r, s)}$ a field with left and right dimensions $(\Delta_{(r, s)}, \Delta_{(r, -s)})$.

The program exposes a Field struct and convenient constructors:

BootstrapVirasoro.Field — Type

Field{T}

Object representing a conformal field. Contains the conformal dimensions, and flags saying whether the field has rational Kac indices and/or is diagonal. The user can access different parametrisations of the left and right conformal dimensions, and Kac indices.

Examples

julia> c = CentralCharge(B = 0.5)
c = 27.999999999999996 + 0.0im, β = 0.0 - 0.7071067811865476im

julia> V = Field(c, r=0, s=2//3)
Non-diagonal Field{ComplexF64}
left: ConformalDimension{ComplexF64} with Kac indices r = 0//1, s = 2//3
right: ConformalDimension{ComplexF64} with Kac indices r = 0//1, s = -2//3

julia> V.r
0//1

julia> V.s
2//3

julia> V.δ
(-0.22222222222222215 - 0.0im, -0.22222222222222215 + 0.0im)

julia> V.δ[:left]
-0.22222222222222215 - 0.0im

julia> V.p[:right]
0.4714045207910316 + 0.0im

julia> V2 = Field(c, δ = 0.5, diagonal=true)
Diagonal Field{ComplexF64} with ConformalDimension{ComplexF64} with
Δ = 1.625 + 0.0im, P = 0.7071067811865476

BootstrapVirasoro.Field — Method

    Field(charge, parameter = value; kwargs)
    Field(charge, r = r, s = s)
    Field((dim_left, dim_right))
    Field(dim_left, dim_right)
    Field(dim) # diagonal field

Constructor function for the Field type.

Given a charge charge, one of the four parameters Δ, δ, P, p and two values, create an object Field{T} (where T is the type of the values in charge) that represents a field of left and right dimensions given by leftvalue and rightvalue in the chosen parametrisation.

keyword arguments:

r::Rational,s::Rational: used in conjunction to Kac=true, must be given rational

values. By convention $V_(r,s)$ has left and right momenta $(P_(r,s), P_(r,-s))$.

diagonal::Bool: set to true to get a diagonal field;

Examples

julia> setprecision(BigFloat, 20, base=10);

julia> c = CentralCharge(β = big"0.5");

julia> V = Field(c, r=0, s=1)
Non-diagonal Field{Complex{BigFloat}}
left: ConformalDimension{Complex{BigFloat}} with Kac indices r = 0, s = 1
right: ConformalDimension{Complex{BigFloat}} with Kac indices r = 0, s = -1

julia> V.Δ
(0.4375 + 0.0im, 0.4375 + 0.0im)

julia> V.P[:left]
1.0

julia> V.p[:right]
-0.0 + 1.0im

julia> V2 = Field(c, :P, 0.42, diagonal=true)
Diagonal Field{Complex{BigFloat}} with ConformalDimension{Complex{BigFloat}} with
Δ = -0.3861000000000000130545 + 0.0im, P = 0.4199999999999999844569

julia> V3 = Field(c, degenerate=true, r=4, s=3//4)
Diagonal Field{Complex{BigFloat}} with ConformalDimension{Complex{BigFloat}} with Kac indices r = 4//1, s = 3//4

Spectra

TODO